Feature/#03/main_page_function
페이지 이동 기능 추가하였습니다
